Refactor SafeZone trait to use CMemoryPoolZone.
