+++ nativelib/src/main/scala-3/scala/scalanative/safe/SafeZone.scala
@@ -1,50 +1,58 @@
 package scala.scalanative.safe
 
+import language.experimental.captureChecking
 import scalanative.unsigned._
 import scala.annotation.implicitNotFound
-import scala.scalanative.unsafe.Tag
+import scala.scalanative.runtime.{RawPtr, CMemoryPool, CMemoryPoolZone}
 
 @implicitNotFound("Given method requires an implicit zone.")
 trait SafeZone {
-
-  /** Return an object of type T allocated in zone. T can be primitive types,
-   *  struct, class.
-   *
-   *  ```
-   *  Expected usage:
-   *  	1. T is primitive type, e.g. Int
-   *  		val x: Int = sz.alloc[Int](10)
-   *  	2. T is struct, e.g. @struct class A(v0: Int, v1: Long) {}
-   *  		val x: A = sz.alloc[A](0, 0L)
-   *  	3. T is class. e.g. class A(v0: Int, v1: Long) {}
-   *  		val x: A = sz.alloc[A](0, 0L)
-   *  		Note that Array[_] is a special class.
-   *  		val x: Array[Int] = sz.alloc[Array[Int]](10) // x.length = 10
-   *  ```
-   *  Currently it's a mock interface which doesn't accept constructor
-   *  parameters.
-   */
-  // def alloc[T]()(using tag: Tag[T]): T
 
   /** Frees allocations. This zone allocator is not reusable once closed. */
   def close(): Unit
 
   /** Return this zone allocator is open or not. */
-  def isOpen: Boolean = !isClosed
+  def isOpen: Boolean
 
   /** Return this zone allocator is closed or not. */
-  def isClosed: Boolean
+  def isClosed: Boolean = !isOpen
+
+  /** Return the handle of this zone allocator. */
+  def handle: RawPtr
 }
+
+final class MemorySafeZone (private[this] val zoneHandle: RawPtr) extends SafeZone {
+
+  private[this] var flagIsOpen = true
+
+  protected def checkOpen(): Unit = {
+    if (!isOpen)
+      throw new IllegalStateException(s"Zone ${this} is already closed.")
+  }
+
+  override def close(): Unit = {
+    checkOpen()
+    flagIsOpen = false
+    CMemoryPoolZone.close(zoneHandle)
+    CMemoryPoolZone.free(zoneHandle)
+  }
+
+  override def isOpen: Boolean = flagIsOpen
+
+  override def handle: RawPtr = zoneHandle
+}
+
 
 object SafeZone {
 
   /** Run given function with a fresh zone and destroy it afterwards. */
-  final def apply[T](f: SafeZone => T): T = {
-    val safeZone = open()
-    try f(safeZone)
-    finally safeZone.close()
+  final def apply[T](f: ({*} SafeZone) => T): T = {
+    val sz: {*} SafeZone = open()
+    try f(sz)
+    finally sz.close()
   }
 
-  final def open(): SafeZone =
-    MemoryPoolSafeZone.open(MemoryPool.defaultMemoryPoolHandle)
+  final def open(): {*} SafeZone = new MemorySafeZone(
+    CMemoryPoolZone.open(CMemoryPool.defaultMemoryPoolHandle)
+  )
 }
